+++ modio/client.py
@@ -6,11 +6,13 @@ import asyncio
 import datetime
 import logging
 import time
+from typing import Optional
 import aiohttp
 import requests
 
 from .errors import modioException
 from .entities import Event, Message, ModFile, Rating, User
+from .enums import TargetPlatform, TargetPortal
 from .objects import Pagination, Returned, Filter
 from .game import Game
 from .mod import Mod
@@ -19,12 +21,14 @@ from .mod import Mod
 class Connection:
     """Class handling under the hood requests and ratelimits."""
 
-    def __init__(self, api_key, access_token, lang, version, test):
+    def __init__(self, api_key, access_token, lang, version, test, platform, portal):
         self.test = test
         self.version = version
         self.access_token = access_token
         self.api_key = api_key
         self.lang = lang
+        self.platform = platform
+        self.portal = portal
 
         self.rate_limit = None
         self.rate_remain = None
@@ -115,6 +119,12 @@ class Connection:
                 "Accept-Language": self.lang,
                 "Content-Type": "application/x-www-form-urlencoded",
             }
+
+        if self.platform is not None:
+            headers["X-Modio-Platform"] = self.platform.value
+
+        if self.portal is not None:
+            headers["X-Modio-Portal"] = self.portal.value
 
         return headers
 
@@ -252,6 +262,10 @@ class Client:
     version : Optional[str]
         An optional keyword argument to allow you to pick a specific version of the API to query,
         usually you shouldn't need to change this. Default is the latest supported version.
+    platform : Optiona[TargetPlatform]
+        The platform to target with requests.
+    portal : Optional[TargetPortal]
+        The portal to target with requests.
 
     Attributes
     -----------
@@ -265,12 +279,18 @@ class Client:
         Is 0 until the rate_remain is 0 and becomes 0 again once the rate limit is reset.
     """
 
-    def __init__(self, *, api_key=None, access_token=None, lang="en", version="v1", test=False):
+    def __init__(self, *, api_key=None, access_token=None, lang="en", version="v1", test=False, platform=None, portal=None):
         self.lang = lang
         self.version = version
         self.test = test
         self.connection = Connection(
-            test=test, api_key=api_key, access_token=access_token, version=version, lang=lang
+            test=test,
+            api_key=api_key,
+            access_token=access_token,
+            version=version,
+            lang=lang,
+            platform=platform,
+            portal=portal,
         )
 
     def __repr__(self):
@@ -287,6 +307,28 @@ class Client:
     @property
     def retry_after(self):
         return self.connection.retry_after
+
+    def set_platform(self, platform: Optional[TargetPlatform] = None) -> None:
+        """Change the platform targetted by the client. Call without an argument to not target any specific
+        paltform.
+
+        Parameters
+        -----------
+        platform : Optional[TargetPlatform]
+            The platform to set
+        """
+        self.connection.platform = platform
+
+    def set_portal(self, portal: Optional[TargetPortal] = None) -> None:
+        """Change the portal targetted by the client. Call without an argument to not target any specific
+        portal.
+
+        Parameters
+        -----------
+        portal : Optional[TargetPortal]
+            The portal to set
+        """
+        self.connection.portal = portal
 
     async def close(self):
         """|async| This function is used to clean up the client in order to close the application that it uses gracefully.